+++ public/nx/markdown/examples/ed-tech.md
@@ -1,0 +1,11 @@
+---
+order: 10
+slug: /examples/ed-tech
+title: Ed Tech
+description: learning
+icon: ed-tech
+image: https://live.staticflickr.com/65535/55043282654_a49615d1df_b.jpg
+tags: examples, ed-tech, learning,
+---
+
+The error is fixed. The Nav component is now a Client Component, and event handlers are only used where allowed. You can now use this navigation without the runtime error in Next.js 1
