+++ public/nx/markdown/examples/index.md
@@ -2,7 +2,7 @@
 order: 100
 slug: /examples
 title: Examples
-description: See us socially
+description: Apps in the wild
 icon: examples
 image: https://live.staticflickr.com/65535/55043282654_a49615d1df_b.jpg
 tags: examples
